reorganize utils - first step
