+++ pylabianca/__init__.py
@@ -1,6 +1,9 @@
 from . import (io, utils, decoding, localize, selectivity, spikes,
-               spike_distance, stats, viz)  # NOQA
+               analysis, spike_distance, stats, viz)  # NOQA
 
+from .analysis import (
+    aggregate, dict_to_xarray, xarray_to_dict
+)  # NOQA
 from .viz import plot_shaded, plot_raster, plot_spikes  # NOQA
 from .spikes import SpikeEpochs, Spikes  # NOQA
 from .selectivity import depth_of_selectivity, explained_variance  # NOQA
